+++ src/main/java/net/lmor/botanicalextramachinery/blocks/tesr/mechanicalOrechid/RenderOrechidAdvanced.java
@@ -17,7 +17,7 @@ public class RenderOrechidAdvanced implements BlockEntityRenderer<BlockEntityOre
     private void renderState(@Nullable BlockState state, float translateX, float translateZ, @Nonnull PoseStack poseStack, @Nonnull MultiBufferSource buffer, int light, int overlay) {
         if (state != null) {
             poseStack.pushPose();
-            poseStack.translate((double)translateX, 0.0, (double)translateZ);
+            poseStack.translate(translateX, 0.0, translateZ);
             Minecraft.getInstance().getBlockRenderer().renderSingleBlock(state, poseStack, buffer, light, overlay);
             poseStack.popPose();
         }
